Add anybar patch
This patch allows any other status bar to replace dwm's status bar. If usealtbar
is set to 1 (true), dwm will not draw or update its own status bar, but instead
attempt to map an alternate status bar with the specified altbarclass class
name. Dwm will for all intents and purposes treat this bar as it is its own
status bar, but will not update it or draw over it.

MONCONTAINS is added as an analog to the INTERSECT macro, but checks within the
full bounds of the monitor rather than just its window area. Using INTERSECT
does not work for bars.