Support altbar if dwm starts after bar
If dwm is restarted, the bar would be treated like a regular window due to the
scan function managing it as a regular window. This commit checks for the bar
class and adds it as a bar in scan(), rather than a regular window.

lrpad and bh are also set to 0 at the start if usealtbar is true, so there are
no gaps while the bar is not mapped.