Set bh to managed bar height
This is to fix an incompatability with the warp patch using a polybar-dwm-module
where clicking on the bar tags would cause a warp since the warp patch would not
take into account m->bh, but only bh when deciding whether or not to warp.

Hopefully this will also squash other potential issues with other patches.

Related to issue mihirlad55/polybar#5.