Add fix for polybar tray
Polybar's tray is a separate window from the bar with override-redirect enabled.
The tray also doesn't send maprequest events and populates the tray somewhat
slowly. This makes managing the polybar tray very cumbersome.

The manage/unmanage functions setup the tray with dwm similarly to the
manage/unmanage functions for the bar.

The destroynotify and unmapnotify also similarly unmanage the tray.

The problem is with determining when to manage the tray since the tray does not
trigger maprequest events and also takes time to finalize its width and x-coord.

Since it does not raise maprequest events, the tray must be scanned for by
checking the class hints for each window and finding the window id of the tray.

Managing the tray too early results in retrieving the width and x-coords too
early which tend to resemble a 1-pixel strip. The next best time to find and
manage the tray is in togglebar. It is somewhat hackish and not a perfect
solution, but it works. The Polybar team does seem to be working on integrating
the tray into the bar. Until then, this fix will be necessary.